publication

+++ src/posts/posts.service.ts
@@ -1,6 +1,6 @@
 import { createPostDto } from './posts.dto';
 import { PostsRepository } from './posts.respository';
-import { Injectable } from '@nestjs/common';
+import { Injectable, HttpException, HttpStatus } from '@nestjs/common';
 
 @Injectable()
 export class PostsService {
@@ -15,15 +15,22 @@ export class PostsService {
     return this.repository.getPosts();
   }
 
-  getPostByParam(id: number) {
+  async getPostByParam(id: number) {
+    const checkPost = await this.repository.getPostByParam(id);
+    if (!checkPost)  throw new HttpException("Not found", HttpStatus.NOT_FOUND);
     return this.repository.getPostByParam(id);
   }
 
-  updatePost(id: number, data: createPostDto) {
+  async updatePost(id: number, data: createPostDto) {
+    const checkPost = await this.repository.getPostByParam(id);
+    if (!checkPost)  throw new HttpException("Not found", HttpStatus.NOT_FOUND);
     return this.repository.updatePost(id, data);
   }
 
-  deletePost(id: number) {
+  async deletePost(id: number) {
+    const checkPost = await this.repository.findPostWithPublications(id);
+    if (!checkPost) throw new HttpException("Not found", HttpStatus.NOT_FOUND);
+    if (checkPost.publications.length) throw new HttpException("Post in use!", HttpStatus.FORBIDDEN);
     return this.repository.deletePost(id);
   }
 }
